Up to 100x faster streaming JSON parsing
Replaces the partial JSON parser with a streaming JSON parser.

```
 ✓ core/util/parser-comparison.bench.ts (2) 5192ms
     name                                hz     min     max    mean     p75     p99    p995    p999     rme  samples
   · parsePartialJson - 5306 tokens  3.7921  261.20  268.88  263.70  264.90  268.88  268.88  268.88  ±0.72%       10
   · streaming parser - 5306 tokens  391.23  2.3913  3.3182  2.5560  2.6857  3.0767  3.3182  3.3182  ±0.82%      196   fastest
```

This parser has been used in production for ~6 months and is extensively tested,
though the API would likely benefit from some changes to better support the
Vercel AI SDK.

There are a few benefits to this parser that should be implemented as a fast
follow.

First, functionally, the AI SDK could offer a stream of patches rather than a
stream of objects. This is exceptionally useful when the stream is processed
server-side, and the stream is intended to pass only a subset of fields through
to a client or to perform pre-processing. These ParseEvents can be used to
efficiently determine which paths are modified.

Second, the parser is more efficient than the current parser in time and space,
and there are ways this could be improved.

1. The work of the parser is O(n) over the input stream, rather than O(n^2) over
   the input stream. A 1000 token stream will parse 1000 tokens with the
   `StreamingParser` from `fn-stream`, rather than ~1.5 million tokens
   `parsePartialJson` must parse over the same stream. (`parsePartialJson` parses
   each prefix thrice.)
2. It is no longer necessary for the server to maintain the accumulated chunks
   or previously parsed objects. This reduces memory demands on the server by
   O(n).
3. A future commit could remove the call to `structuredClone` and reduce the
   number of copies of the object again. The tests should be updated to simulate
   a serialize-deserialize operation and introduce asynchrony. The tests
   currently provide all of the chunks synchronously to `streamObject` and as the
   tests simulate a client and server in the same process, the enqueued object is
   shared, so the assertions only see the final state. This would remove one more
   copy, which becomes accidentally quadratic over the stream.