Fix unsafe use of `any` in `stream-object.ts`
The stream transformer used in `streamObject` permitted chunks of type `any`,
which was then cast to a stream of `string | ObjectStreamInputPart`. However
this was not type safe. The `usage` key was incorrect. This was papered over by
assigning the `usage = calculateTokenUsage(chunk.usage)`, but the types also
permitted `usage = chunk.usage`.

This fixes `streamObject` to generate a correctly typed stream passed to the
`StreamObjectResult` constructor.