fix (provider/bedrock): don't use env var for session token default.
Deployments using the Bedrock provider can run in a serverless
function that pre-declares a value for `AWS_SESSION_TOKEN`
representing an alternate set of credentials than those intended and
specified by the developer. If we silently use the environment
variable for a default it can create a non-obvious credentials failure
that surprises the developer.

The session token credentials are less commonly used, and the
developer can still pass them along manually (and reference the
environment variable themselves if they choose to do so).

We also update the documentation to state this behavior as it's
admittedly deviant from common practice in the AI SDK providers (but
for good reason due to the idiosyncratic way it can be already-present
in the serverless environment).

Fixes #3018.