core (ai/mcp): update experimental MCP client documentation for Streamable HTTP transport usage (#5972)
The current experimental MCP client already supports the new HTTP
transport through the custom transport option (over passing in a
config).

Updates documentation to illustrate usage of the new Streamable HTTP
Transport with `experimental_createMCPClient` for tool conversion.

As of now, `experimental_createMCPClient` is only used in the AI SDK as
a way to fetch and call MCP server tools; for this reason, we are
**not** adding support for Session Management or Resumable Streams -
which are new features supported by the new HTTP transport.

Examples have been updated to illustrate basic usage, and a new Next.js
example has been added to illustrate setting up a **stateless** MCP
**server** to be used with `useChat` and `streamText`.

<!--
This task list is intended to help you keep track of what you need to
do.
Feel free to add tasks and remove unnecessary tasks as needed.

Please check if the PR fulfills the following requirements:
-->

- [x] Tests have been added / updated (for bug fixes / features)
- [x] Documentation has been added / updated (for bug fixes / features)
- [ ] A _patch_ changeset for relevant packages has been added (for bug
fixes / features - run `pnpm changeset` in the project root)
- [x] Formatting issues have been fixed (run `pnpm prettier-fix` in the
project root)

Given the deprecation of the SSE transport, we should likely deprecate
on our end as well. We should consider fully deprecating native support
for transports (e.g. our custom `stdio` transport), and enforcing users
to always pass in a custom transport to limit need for maintenance on
our end.

Fixes #5984