feat (provider): add `providerMetadata` to `ImageModelV2` interface (#5977)
## Background

Compare #5698

## Summary

Additional provider-specific options to the image model provider
interface. They are passed through to the provider from the AI SDK and
enable provider-specific functionality that can be fully encapsulated in
the provider.

Unlike other models, ImageModel request return an array of images, and
provider can return image-specific metadata for each. So far, this pull
request passing through the revised prompt used for each image. In order
to make that possible, I introduced a new type
`ImageModelV2ProviderMetadata` which is the same as
`SharedV2ProviderMetadata` plus it guarantees the presence of the
`.images` key

```js
export type ImageModelV2ProviderMetadata = Record<
  string,
  {
    images: JSONArray;
  } & JSONValue
>;
```

That also makes it possible to deeply merge providerMetadata from
multiple responses effectively.

## Verification

I updated the `examples/ai-core/src/generate-image/openai.ts` example to
verify that the code is working.