fix(providers/amazon-bedrock): strip empty text content to avoid API errors.
Without this, with maxSteps set, we sometimes see messages sent to the API like this:

```
{
  role: 'assistant',
  content: [ { text: '\n\n' }, { toolUse: [Object] } ]
},
```

This leads to API errors like this:

```
The text field in the ContentBlock object at messages.$n.content.0 is blank. Add text to the text field, and try again.
```

This change simply filters out this useless but apparently harmful message.