feat (ai): replace maxSteps with continueUntil (generateText) (#6338)
## Background

The `maxSteps` parameter is not flexible enough for use cases that
require more fine grained controlled over when the steps loop should
end, e.g. when token consumption is exhausted or when a specific tool
has been called.

## Summary

Replace `maxSteps` in `generateText` with `continueUntil`, which takes a
more flexible `StopCondition`.

## Future Work

- Implement `continueUntil` for `streamText`
- Array support for `continueUntil`