feat (ai): replace maxSteps with continueUntil (streamText) (#6393)
## Background

The `maxSteps` parameter is not flexible enough for use cases that
require more fine grained controlled over when the steps loop should
end, e.g. when token consumption is exhausted or when a specific tool
has been called.

## Summary

Replace `maxSteps` in `streamText` with `continueUntil`, which takes a
more flexible `StopCondition`.

## Future Work

- Array support for `continueUntil`