feat (ai): add array support to stopWhen (#6339)
## Background

When an additional stop condition such as `hasToolCall` is used, it
makes sense to still keep a step limit condition in case the additional
condition is never met to prevent infinite loops.

## Summary

Add support for multiple stop conditions passed as an array with `or`
semantics. When any stop condition is met, the multi-step loop is
aborted.

## Tasks

- [x] implementation
- [x] add tests
- [x] changeset update
- [x] jsdoc
- [x] extract