feat (providers/gateway): include deployment and request id (#6591)
## Background

For o11y it is useful to have a developer's deployment and request id.
The deployment id variable name was incorrect, and we don't yet include
logic to look up the request id.

## Summary

Corrected the environment variable name to look up the deployment id,
and added logic to look for the request id in the headers as
`x-vercel-id`.

## Verification

Added verbose logging and deployed to a preview deployment to observe
correct data being logged.

## Tasks

- [x] Tests have been added / updated (for bug fixes / features)
- [ ] Documentation has been added / updated (for bug fixes / features)
- [x] A _patch_ changeset for relevant packages has been added (for bug
fixes / features - run `pnpm changeset` in the project root)
- [x] Formatting issues have been fixed (run `pnpm prettier-fix` in the
project root)