feat (ai): add args callbacks to tools (#6681)
## Background

In order to observe tool calls in progress and to send data parts to the
client, it is currently necessary to process the `fullStream` output or
to rely on UI message tool invocation parts. This change will enable
responding to args streaming events in the context of a particular tool.

## Summary

* rename `ToolExectionOptions` to `ToolCallOptions`
* add `onArgsStreamingStart`, `onArgsStreamingDelta`, and
`onArgsAvailable` callbacks to tools.

## Verification

* verify that they are called in next js route example

## Tasks

- [x] onArgsAvailable streamText
- [x] onArgsAvailable generateText
- [x] onArgsStreamingStart streamText
- [x] onArgsStreamingDelta streamText
- [x] changeset