chore: use Zod 4 internally (#6884)
## Background

Moving to a newer version of Zod for internal types would make it easier
for 3rd party clients to still just use either the server or UI stuff.

## Summary

Mostly just changing imports, and a few things here/there to support for
that, mostly based on migration guide from Zod team.