removes additionalProperties from req to GoogleAPI
Google APIs support only a subset of OpenAPI schema.
Ref: https://ai.google.dev/api/caching#Schema

And `additionalProperties` is not one of them. Passing this property in
the request leads to 400 BAD REQUEST error from Google APIs, complaining
specifically about the invalid parameter: `additionalProperties`.

In the following PR (https://github.com/vercel/ai/pull/6919) we added an
important fix to identify properties incorrectly getting converted to
undefined because they did not have `properties` field but we also
introduced `additionalProperties` in the converted schema.

This commit keeps the change in `isEmptyObjectSchema` for identifying
the objects with empty or no `properties` field and reverts the
introduction of `additionalProperties`.