Add Vercel SDK Provider list (#6863)
<!--
Welcome to contributing to AI SDK! We're excited to see your changes.

We suggest you read the following contributing guide we've created
before submitting:

https://github.com/vercel/ai/blob/main/CONTRIBUTING.md
-->

## Background

<!-- Why was this change necessary? -->

I noticed that in the [Providers and Models
docs](https://ai-sdk.dev/docs/foundations/providers-and-models), we
mention the Vercel (v0) model in the [Model
Capabilities](https://ai-sdk.dev/docs/foundations/providers-and-models#model-capabilities)
but not in the [AI SDK
Providers](https://ai-sdk.dev/docs/foundations/providers-and-models#ai-sdk-providers).

## Summary

<!-- What did you change? -->

This PR adds the Vercel Provider to the [Providers and Models
docs](https://ai-sdk.dev/docs/foundations/providers-and-models#ai-sdk-providers)
list.

## Tasks

<!--
This task list is intended to help you keep track of what you need to
do.
Feel free to add tasks and remove unnecessary tasks as needed.

Please check if the PR fulfills the following requirements:
-->

- [x] Formatting issues have been fixed (run `pnpm prettier-fix` in the
project root)