Add Vercel SDK Provider list (#6863) (#6961)
## Background

I noticed that in the [Providers and Models
docs](https://ai-sdk.dev/docs/foundations/providers-and-models), we
mention the Vercel (v0) model in the [Model

Capabilities](https://ai-sdk.dev/docs/foundations/providers-and-models#model-capabilities)
but not in the [AI SDK

Providers](https://ai-sdk.dev/docs/foundations/providers-and-models#ai-sdk-providers).

## Summary

This PR adds the Vercel Provider to the [Providers and Models
docs](https://ai-sdk.dev/docs/foundations/providers-and-models#ai-sdk-providers)
list.

## Tasks

- [x] Formatting issues have been fixed (run `pnpm prettier-fix` in the
project root)

Co-authored-by: Anshuman Bhardwaj <anshu5074@gmail.com>