feat (providers/gateway): include description and pricing info in model list (#7081)
## Background

Vercel AI Gateway customers would like programmatic access to pricing
info and brief description for the models in the model library.

## Summary

Add input/output token cost for the model entries produced by
`getAvailableModels` in the Gateway provider.

## Verification

Updated unit tests and tested manually with debug output in an example
script.

## Future Work

We aren't using the `specification` data in `GatewayLanguageModelEntry`
type. I looked at removing it as part of this change as dead code/data,
but as we've exported the specification type in the provider's
`index.ts`, removing it wholesale would be a breaking change. Leaving
further work here for the future.