feat (ai): add readUIMessageStream helper (#7116)
## Background

UI message streams are useful outside of `Chat` use cases, e.g. for
terminal UIs, custom stream consumption on the client, or RSC.

## Summary

Add a `readMessageStream` helper that consumes a UI message stream
consisting of `UIMessageChunk` objects and returns and async iterable
stream over the same `UIMessage` as it is being constructed.

## Validation

Added and example and ran it.

---------

Co-authored-by: nicoalbanese <gcalbanese96@gmail.com>