docs(CONTRIBUTING): add codemods instructions (#7085)
## Background

As part of [the AI SDK
v5](https://ai-sdk.dev/docs/announcing-ai-sdk-5-beta), we create a lot
of codemods to streamline the upgrade process for existing users.
Instead of waiting until another breaking change release, we want to
establish a process of creating codemods as part of pull requests, where
applicable. Pull requests that should attempt to add a codemod usually
introduce a deprecation or a breaking change.

## Summary

Add new `How to create a codemod` section to `CONTRIBUTING.md`

## Tasks


- [x] Documentation has been added / updated (for bug fixes / features)
- [x] Formatting issues have been fixed (run `pnpm prettier-fix` in the
project root)

---------

Co-authored-by: josh <144584931+dancer@users.noreply.github.com>