fix(@ai-sdk/google): Use single embeddings endpoint when using embed() (#7328)
From #7243 :

## Background

We use the `@ai-sdk/google` package at supermemory, and many times, use
the `embed()` functionality to generate embeddings for single queries
etc.

However, I noticed that the AI sdk was _always_ calling the batch embed
content, and we frequently ran into ratelimits because that endpoint has
a limit of 150, as opposed to 1500 for the single `:embedContent`
endpoint. The ratelimit is measured by number of requests, not number of
items in the batch for google.

This PR makes sure that when there's `embed` (a single embedding), the
request goes through the `embedContent` endpoint, ensuring that the
right type of ratelimit is enforced. It also reads the result and
formats it in the same output format as expected.

## Summary

I added a function `doEmbedSingle` that uses the single `embedContent`
URL `${this.config.baseURL}/models/${this.modelId}:embedContent`

## Verification

I temporarily added a console.log in the doEmbedSingle, ran it, and also
verified that it does not break by running the batch one again.

I ran the `pnpm tsx src/embed/google.ts` with my
`GOOGLE_GENERATIVE_AI_API_KEY` and confirmed that it works.

<img width="1024" height="532" alt="image"
src="https://github.com/user-attachments/assets/f3b3f859-fb8f-4db4-891c-1b8b922a093f"
/>

## Related Issues

Continues #7243

---------

Co-authored-by: Dhravya Shah <dhravyashah@gmail.com>