feat (ai): add isAborted to onFinish callback for ui message streams (#7400)
## Background

We recommend persisting UI messages. However, the `streamText.onAbort`
callback works on the model message level.

## Summary

* send abort chunks in the ui message stream
* introduce `UIMessageStreamOnFinishCallback` type
* add `isAborted` flag to `UIMessageStreamOnFinishCallback`
* add response aborted error type for nextjs
* expose consumeStream

## Note

Consuming the UI message stream is necessary because the client/server
connection is gone when the fetch is aborted from the client, and
because of the backpressure we need a separate ui message stream
consumer.

## Verification

- [x] Check next-openai abort scenario manually

## Related Issues

Continues work from #7394

---------

Co-authored-by: nicoalbanese <gcalbanese96@gmail.com>