fix (provider/gateway): remove unnecessary `x-` prefix on auth method header (#7407)
## Background

The Gateway provider added a header to report the auth method used to
the server. We included an `x-` prefix but this is unnecessary and makes
the header different from our other existing `ai-` headers.

## Summary

Standardize on `ai-` as a prefix and remove the `x-` prefix from the
auth method header.