feat (docs): add built-in-ai custom provider (#7320)
## Background

This PR adds `built-in-ai` as a community provider that utilizes the
browser's native JavaScript [Prompt
API](https://github.com/webmachinelearning/prompt-api) to provide access
to native browser AI models in Chrome & Edge with support for the Vercel
AI SDK.

It will also serves as the AI SDK provider for other client side
in-browser AI models (libraries), meaning we'll aim to add support for
'web-llm' and 'transformers.js' under the @built-in-ai org.

While there is an [existing
package](https://github.com/jeasonstudio/chrome-ai) that attempted to
provide similar functionality for Chrome, it has been inactive for the
past year with no maintenance or updates. It's not updated to work with
the Prompt API nor the AI SDK.

There is [an issue](https://github.com/jeasonstudio/chrome-ai/issues/32)
created in December last year, asking about the status of the repository
that received no response.
The open issue with no responses and a year of inactivity, suggests that
the package will not be maintained.

## Summary

I added the Built-in AI custom provider to the
`content/providers/03-community-providers` and updated the
`content\docs\02-foundations\02-providers-and-models.mdx`.

## Future Work

When the Vercel AI SDK v5 is released to stable, the '<Note>' components
with warnings should be removed from the documentation .mdx