feat (ui): Add ability to "reset" to structured object (#7272)
## Background

I'm building an application, and I want a way to clear the existing
object without reinitialising the StructuredObject class or submitting
the object.

This is a wanted feature if the thumbs up on this [GH
issue](https://github.com/vercel/ai/issues/2397) are any indication.

For me, I'm using `StructuredObject` instead of `Chat` to manage a
conversation. It means that I'm manually managing the messages logic.

Once the object finishes streaming, I want to clear it and append it to
the list of messages.

## Summary

Added a new field to the object that allows us to reset the object.

I also spotted a bug where the error created via the AbortController was
being returned falsy when pass through to `isAbortError`