feat(anthropic): Add raw usage object to Anthropic provider metadata (#7591)
## Background

Anthropic supports several different flavors of prompt caching in beta.
These are priced differently. It would be lovely to give tracing
providers a way to accurately track this.

## Summary

Adds the raw returned Anthropic `usage` object to provider metadata. A
subset of this is returned as
`providerMetadata.anthropic.cache_creation_input_tokens`, but this is
not enough granularity to tell whether it was a 5 minute cache or a 1 hr
cache.

## Verification

See added example.

## Related Issues
Continues #7566

---------

Co-authored-by: jacoblee93 <jacoblee93@gmail.com>