feat (ai): add static/dynamic toolCalls/toolResults helpers (#7617)
## Background

The introduction of dynamic tools in #7595 broke type inference in
several scenarios (see #7615 ).

## Summary

Given that we want to keep both dynamic tools calls/results and static
tools calls/results in a single calls/results array, the type inference
limitations cannot be resolved directly. However, it is possible to
expose subsets for static/dynamic calls/results that have the expected
typing. This has the advantage that it makes it explicit that you are
only accessing static or dynamic calls/results.

* renamed `ToolCallUnion` to `TypedToolCall`
* expose `StaticToolCall` and `DynamicToolCall` types
* renamed `ToolResultUnion` to `TypedToolResult`
* expose `StaticToolResult` and `DynamicToolResult` types
* expose `StaticToolError`, `DynamicToolError`, `TypedToolError` types
* add `staticToolCall` and `dynamicToolCalls` helpers to `generateText`,
`streamText`, `Step`
* add `staticToolResult` and `dynamicToolResult` helpers to
`generateText`, `streamText`, `Step`

## Verification

Verified type inference works as before when switching to `static*`
helpers.

## Related Issues

Resolves #7615