Set status to ready when reconnect stream is null
In the resume-stream flow, an active stream may not exist. In this case, the server returns 204 and the client should move on. However, prior to this fix, the useChat status would remain as "submitted" instead of being reset to "ready". This commit addresses this issue by setting status to "ready" before returning early in the makeRequest function.